Copy files incrementally only
In case of a lot of and/or large assets copying all files over and over
results in timely and needless file I/O.
This change will compare file sizes and mtimes of source and
destination and skip identical files.
It also introduces a plugin option 'forceCopy' (defaults to `false`)
which results in the old behavior.